Revert "improved error handling, fixed several warnings, started to introduce google style code format"
This reverts commit 1d7ae4a7

+++ src/main/java/com/pj/squashrestapp/dto/match/MatchesSimplePaginated.java
@@ -30,10 +30,10 @@ public class MatchesSimplePaginated {
     this.total = page.getTotalElements();
     this.page = page.getNumber();
     this.pages = page.getTotalPages();
-    this.min = (long) page.getNumber() * page.getSize() + 1;
+    this.min = page.getNumber() * page.getSize() + 1;
     this.max = (page.getNumber() == page.getTotalPages() - 1)
             ? page.getTotalElements()
-            : (long) (page.getNumber() + 1) * matchesDtos.size();
+            : (page.getNumber() + 1) * matchesDtos.size();
     this.matches = getSortedMatches(matchesDtos);
   }
